I noticed a limitation of my design during final testing. The sales price of each trip was not being updated after the save of a new service!
Since this is logic which is implmeneted on the server, it requires us to query the dataservice after a service is saved for its corresponding trip.

This required changing the trip view to inherit from the class GenericEndpoint class I made prior for Service and Category. Now the primary key can be passed to the dataservice to query for a specific trip.

Also updated the front end to execute this query after a successful save of a new service.